feat(ui): split cron jobs and routines into separate routed pages; move REST API under /api/v1
The web UI previously rendered both cron jobs and routines as tabs inside a
single view. This introduces real client-side routing (yew-router) so each is
its own addressable page:

  /            -> redirects to /cron-jobs
  /cron-jobs   -> Cron Jobs page
  /routines    -> Routines page

Because the UI is served from the root path space, the REST API is moved under
an /api/v1 prefix so UI paths (/cron-jobs, /routines) no longer collide with the
JSON resource endpoints. The server serves the SPA HTML for any unmatched path
(fallback) so deep links and refreshes resolve in the router.

Server:
- Nest all REST routes under /api/v1 (health, shutdown, echo, cron-jobs*,
  agents, routines*, routines.ics, cleanup).
- Add SPA fallback so /cron-jobs, /routines, etc. return the app HTML.
- OpenAPI server URL carries the /api/v1 base; drop the UI root from the spec.
- Update CLI internal calls (stop/cleanup/health) and startup banner.

UI:
- Extract cron-jobs UI into a self-contained CronJobsPage (cron_jobs.rs),
  mirroring RoutinesPage; main.rs becomes a router shell (header, nav, switch,
  shutdown dialog, toasts).
- Point all fetches at /api/v1.

Docs/tests updated; SPA-fallback test added.
