+++ src/cli.rs
@@ -109,7 +109,7 @@ pub fn run_background() -> anyhow::Result<()> {
 
 /// Ask a running server to stop via the `/shutdown` route.
 pub fn stop() -> anyhow::Result<()> {
-    match http_request("POST", "/shutdown") {
+    match http_request("POST", "/api/v1/shutdown") {
         Ok(200) => {
             println!("moadim is shutting down");
             Ok(())
@@ -129,7 +129,7 @@ pub fn stop() -> anyhow::Result<()> {
 /// Runs the same sweep as the hourly background task instead of waiting for the next tick, via the
 /// `/routines/cleanup` route. Prints how many workbenches were removed, or a hint when no server is up.
 pub fn cleanup() -> anyhow::Result<()> {
-    match http_request_with_body("POST", "/routines/cleanup") {
+    match http_request_with_body("POST", "/api/v1/routines/cleanup") {
         Ok((200, body)) => {
             let removed = parse_removed_count(&body).unwrap_or(0);
             let plural = if removed == 1 { "" } else { "es" };
@@ -201,7 +201,7 @@ fn paths_daemon_log() -> String {
 
 /// Returns `true` if a server answers `GET /health` on [`BIND_ADDR`].
 pub(crate) fn is_running() -> bool {
-    matches!(http_request("GET", "/health"), Ok(200))
+    matches!(http_request("GET", "/api/v1/health"), Ok(200))
 }
 
 /// Send a minimal HTTP/1.1 request to the local server and return the response status code.
