+++ ui/src/main.rs
@@ -1,32 +1,17 @@
 use croner::Cron;
 use gloo_net::http::Request;
 use gloo_timers::future::TimeoutFuture;
-use serde::{Deserialize, Serialize};
-use serde_json::Value as Json;
+use serde::Deserialize;
 use wasm_bindgen_futures::spawn_local;
-use web_sys::HtmlInputElement;
 use yew::prelude::*;
+use yew_router::prelude::*;
 
+mod cron_jobs;
 mod routines;
+use cron_jobs::CronJobsPage;
 use routines::RoutinesPage;
 
-// ─── Types (mirror server API exactly) ────────────────────────────────────────
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct CronJob {
-    pub id: String,
-    pub schedule: String,
-    pub handler: String,
-    pub metadata: Json,
-    pub enabled: bool,
-    pub created_at: u64,
-    pub updated_at: u64,
-    #[serde(default)]
-    pub last_triggered_at: Option<u64>,
-    /// Human-readable schedule description supplied by the server (e.g. "At 09:30, Monday through Friday").
-    #[serde(default)]
-    pub schedule_description: Option<String>,
-}
+// ─── Shared types ─────────────────────────────────────────────────────────────
 
 #[derive(Debug, Clone, Deserialize, PartialEq, Default)]
 pub struct Health {
@@ -34,28 +19,6 @@ pub struct Health {
     pub uptime_secs: Option<u64>,
     pub running: bool,
 }
-
-#[derive(Debug, Clone, Serialize)]
-pub struct CreateRequest {
-    pub schedule: String,
-    pub handler: String,
-    pub metadata: Json,
-    pub enabled: bool,
-}
-
-#[derive(Debug, Clone, Serialize, Default)]
-pub struct UpdateRequest {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub schedule: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub handler: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub metadata: Option<Json>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub enabled: Option<bool>,
-}
-
-// ─── App state ────────────────────────────────────────────────────────────────
 
 #[derive(Debug, Clone, PartialEq)]
 pub enum ToastKind {
@@ -70,113 +33,54 @@ pub struct Toast {
     pub kind: ToastKind,
 }
 
-#[derive(Debug, Clone, PartialEq, Default)]
-pub enum Page {
-    #[default]
-    List,
-    NewJob,
-    Logs(String),
-}
+// ─── Routes ───────────────────────────────────────────────────────────────────
 
-/// Top-level navigation tab.
-#[derive(Debug, Clone, Copy, PartialEq, Default)]
-pub enum Tab {
-    #[default]
-    Jobs,
+/// Top-level pages, served at the root path space: `CronJobs` at `/cron-jobs` and `Routines` at
+/// `/routines`. The REST API is namespaced under `/api/v1`, so these UI paths never collide with it.
+/// The server returns the same self-contained HTML for any unmatched path (SPA fallback), letting
+/// these deep links and refreshes load the app so the router can resolve the path.
+#[derive(Clone, Routable, PartialEq)]
+pub enum Route {
+    #[at("/")]
+    Home,
+    #[at("/cron-jobs")]
+    CronJobs,
+    #[at("/routines")]
     Routines,
+    #[not_found]
+    #[at("/404")]
+    NotFound,
 }
 
-#[derive(Debug, Clone, PartialEq)]
-pub enum Modal {
-    None,
-    Edit(String),
-    ConfirmDelete { id: String, handler: String },
-    ConfirmShutdown,
-}
+// ─── Shell state (health, toasts, shutdown) ───────────────────────────────────
 
-#[derive(Debug, Clone, PartialEq)]
-pub struct AppState {
-    pub jobs: Vec<CronJob>,
+#[derive(Debug, Clone, PartialEq, Default)]
+pub struct ShellState {
     pub health: Health,
     pub health_ok: bool,
-    pub loading: bool,
-    pub tab: Tab,
-    pub page: Page,
-    pub modal: Modal,
     pub toasts: Vec<Toast>,
     pub next_toast: u32,
+    pub show_shutdown: bool,
 }
 
-impl Default for AppState {
-    fn default() -> Self {
-        Self {
-            jobs: vec![],
-            health: Health::default(),
-            health_ok: false,
-            loading: true,
-            tab: Tab::Jobs,
-            page: Page::List,
-            modal: Modal::None,
-            toasts: vec![],
-            next_toast: 0,
-        }
-    }
-}
-
-pub enum AppAction {
-    SwitchTab(Tab),
-    JobsLoaded(Vec<CronJob>),
+pub enum ShellAction {
     HealthLoaded { health: Health, ok: bool },
-    GoToCreate,
-    GoToList,
-    GoToLogs(String),
-    OpenEdit(String),
-    OpenConfirmDelete { id: String, handler: String },
-    OpenConfirmShutdown,
-    CloseModal,
-    UpsertJob(CronJob),
-    RemoveJob(String),
     AddToast { msg: String, kind: ToastKind },
-    DismissToast(u32),
+    OpenShutdown,
+    CloseShutdown,
 }
 
-impl Reducible for AppState {
-    type Action = AppAction;
+impl Reducible for ShellState {
+    type Action = ShellAction;
 
     fn reduce(self: std::rc::Rc<Self>, action: Self::Action) -> std::rc::Rc<Self> {
         let mut s = (*self).clone();
         match action {
-            AppAction::SwitchTab(tab) => {
-                s.tab = tab;
-                s.page = Page::List;
-                s.modal = Modal::None;
-            }
-            AppAction::JobsLoaded(jobs) => {
-                s.jobs = jobs;
-                s.loading = false;
-            }
-            AppAction::HealthLoaded { health, ok } => {
+            ShellAction::HealthLoaded { health, ok } => {
                 s.health = health;
                 s.health_ok = ok;
             }
-            AppAction::GoToCreate => s.page = Page::NewJob,
-            AppAction::GoToList => s.page = Page::List,
-            AppAction::GoToLogs(id) => s.page = Page::Logs(id),
-            AppAction::OpenEdit(id) => s.modal = Modal::Edit(id),
-            AppAction::OpenConfirmDelete { id, handler } => {
-                s.modal = Modal::ConfirmDelete { id, handler };
-            }
-            AppAction::OpenConfirmShutdown => s.modal = Modal::ConfirmShutdown,
-            AppAction::CloseModal => s.modal = Modal::None,
-            AppAction::UpsertJob(job) => {
-                if let Some(i) = s.jobs.iter().position(|j| j.id == job.id) {
-                    s.jobs[i] = job;
-                } else {
-                    s.jobs.push(job);
-                }
-            }
-            AppAction::RemoveJob(id) => s.jobs.retain(|j| j.id != id),
-            AppAction::AddToast { msg, kind } => {
+            ShellAction::AddToast { msg, kind } => {
                 let id = s.next_toast;
                 s.next_toast += 1;
                 s.toasts.push(Toast {
@@ -185,7 +89,8 @@ impl Reducible for AppState {
                     kind,
                 });
             }
-            AppAction::DismissToast(id) => s.toasts.retain(|t| t.id != id),
+            ShellAction::OpenShutdown => s.show_shutdown = true,
+            ShellAction::CloseShutdown => s.show_shutdown = false,
         }
         s.into()
     }
@@ -193,18 +98,8 @@ impl Reducible for AppState {
 
 // ─── API layer ────────────────────────────────────────────────────────────────
 
-async fn api_list_jobs() -> Result<Vec<CronJob>, String> {
-    Request::get("/cron-jobs")
-        .send()
-        .await
-        .map_err(|e| e.to_string())?
-        .json::<Vec<CronJob>>()
-        .await
-        .map_err(|e| e.to_string())
-}
-
 async fn api_health() -> Result<Health, String> {
-    Request::get("/health")
+    Request::get("/api/v1/health")
         .send()
         .await
         .map_err(|e| e.to_string())?
@@ -213,57 +108,8 @@ async fn api_health() -> Result<Health, String> {
         .map_err(|e| e.to_string())
 }
 
-async fn api_create(req: &CreateRequest) -> Result<CronJob, String> {
-    let resp = Request::post("/cron-jobs")
-        .json(req)
-        .map_err(|e| e.to_string())?
-        .send()
-        .await
-        .map_err(|e| e.to_string())?;
-    if !resp.ok() {
-        return Err(format!("HTTP {}", resp.status()));
-    }
-    resp.json::<CronJob>().await.map_err(|e| e.to_string())
-}
-
-async fn api_update(id: &str, req: &UpdateRequest) -> Result<CronJob, String> {
-    let resp = Request::put(&format!("/cron-jobs/{id}"))
-        .json(req)
-        .map_err(|e| e.to_string())?
-        .send()
-        .await
-        .map_err(|e| e.to_string())?;
-    if !resp.ok() {
-        return Err(format!("HTTP {}", resp.status()));
-    }
-    resp.json::<CronJob>().await.map_err(|e| e.to_string())
-}
-
-async fn api_delete(id: &str) -> Result<(), String> {
-    let resp = Request::delete(&format!("/cron-jobs/{id}"))
-        .send()
-        .await
-        .map_err(|e| e.to_string())?;
-    if resp.status() == 204 || resp.ok() {
-        Ok(())
-    } else {
-        Err(format!("HTTP {}", resp.status()))
-    }
-}
-
-async fn api_trigger(id: &str) -> Result<CronJob, String> {
-    let resp = Request::post(&format!("/cron-jobs/{id}/trigger"))
-        .send()
-        .await
-        .map_err(|e| e.to_string())?;
-    if !resp.ok() {
-        return Err(format!("HTTP {}", resp.status()));
-    }
-    resp.json::<CronJob>().await.map_err(|e| e.to_string())
-}
-
 async fn api_shutdown() -> Result<(), String> {
-    let resp = Request::post("/shutdown")
+    let resp = Request::post("/api/v1/shutdown")
         .send()
         .await
         .map_err(|e| e.to_string())?;
@@ -274,46 +120,52 @@ async fn api_shutdown() -> Result<(), String> {
     }
 }
 
-async fn api_get_logs(id: &str) -> Result<String, String> {
-    let resp = Request::get(&format!("/cron-jobs/{id}/logs"))
-        .send()
-        .await
-        .map_err(|e| e.to_string())?;
-    if !resp.ok() {
-        return Err(format!("HTTP {}", resp.status()));
+async fn poll_health(state: UseReducerHandle<ShellState>) {
+    match api_health().await {
+        Ok(health) => {
+            let ok = health.running;
+            state.dispatch(ShellAction::HealthLoaded { health, ok });
+        }
+        Err(_) => state.dispatch(ShellAction::HealthLoaded {
+            health: Health {
+                status: "offline".into(),
+                running: false,
+                uptime_secs: None,
+            },
+            ok: false,
+        }),
     }
-    resp.text().await.map_err(|e| e.to_string())
 }
 
-// ─── Root component ───────────────────────────────────────────────────────────
+// ─── Root ─────────────────────────────────────────────────────────────────────
 
 #[function_component(App)]
 pub fn app() -> Html {
-    let state = use_reducer(AppState::default);
+    html! {
+        <BrowserRouter>
+            <Shell />
+        </BrowserRouter>
+    }
+}
 
-    // Load jobs + poll health on mount
+/// Persistent chrome around the routed pages: header (health + STOP), nav tabs, the routed page,
+/// the global shutdown dialog, and the toast stack. Lives inside the router so nav `Link`s work.
+#[function_component(Shell)]
+pub fn shell() -> Html {
+    let state = use_reducer(ShellState::default);
+
+    // Initial health poll on mount.
     {
         let state = state.clone();
         use_effect_with((), move |_| {
-            let state = state.clone();
-            spawn_local(async move {
-                match api_list_jobs().await {
-                    Ok(jobs) => state.dispatch(AppAction::JobsLoaded(jobs)),
-                    Err(e) => state.dispatch(AppAction::AddToast {
-                        msg: format!("Failed to load jobs: {e}"),
-                        kind: ToastKind::Err,
-                    }),
-                }
-                poll_health(state).await;
-            });
+            spawn_local(async move { poll_health(state).await });
         });
     }
 
-    // Health poll loop every 30 s
+    // Health poll loop every 30 s.
     {
         let state = state.clone();
         use_effect_with((), move |_| {
-            let state = state.clone();
             spawn_local(async move {
                 loop {
                     TimeoutFuture::new(30_000).await;
@@ -323,39 +175,41 @@ pub fn app() -> Html {
         });
     }
 
+    let on_toast = {
+        let state = state.clone();
+        Callback::from(move |(msg, kind): (String, ToastKind)| {
+            state.dispatch(ShellAction::AddToast { msg, kind })
+        })
+    };
+
     let on_refresh = {
         let state = state.clone();
         Callback::from(move |_: MouseEvent| {
             let state = state.clone();
-            spawn_local(async move {
-                match api_list_jobs().await {
-                    Ok(jobs) => state.dispatch(AppAction::JobsLoaded(jobs)),
-                    Err(e) => state.dispatch(AppAction::AddToast {
-                        msg: format!("Refresh failed: {e}"),
-                        kind: ToastKind::Err,
-                    }),
-                }
-                poll_health(state).await;
-            });
+            spawn_local(async move { poll_health(state).await });
         })
     };
 
-    // The STOP button only opens a confirmation dialog; the server is not asked
-    // to stop until the user confirms (see `on_confirm_shutdown`).
+    // STOP only opens a confirmation dialog; the server is asked to stop on confirm.
     let on_stop = {
         let state = state.clone();
-        Callback::from(move |_: MouseEvent| state.dispatch(AppAction::OpenConfirmShutdown))
+        Callback::from(move |_: MouseEvent| state.dispatch(ShellAction::OpenShutdown))
+    };
+
+    let on_close_shutdown = {
+        let state = state.clone();
+        Callback::from(move |_: ()| state.dispatch(ShellAction::CloseShutdown))
     };
 
     let on_confirm_shutdown = {
         let state = state.clone();
         Callback::from(move |_: ()| {
             let state = state.clone();
-            state.dispatch(AppAction::CloseModal);
+            state.dispatch(ShellAction::CloseShutdown);
             spawn_local(async move {
                 match api_shutdown().await {
                     Ok(()) => {
-                        state.dispatch(AppAction::HealthLoaded {
+                        state.dispatch(ShellAction::HealthLoaded {
                             health: Health {
                                 status: "stopping".into(),
                                 running: false,
@@ -363,12 +217,12 @@ pub fn app() -> Html {
                             },
                             ok: false,
                         });
-                        state.dispatch(AppAction::AddToast {
+                        state.dispatch(ShellAction::AddToast {
                             msg: "Server stopping…".into(),
                             kind: ToastKind::Ok,
                         });
                     }
-                    Err(e) => state.dispatch(AppAction::AddToast {
+                    Err(e) => state.dispatch(ShellAction::AddToast {
                         msg: format!("Stop failed: {e}"),
                         kind: ToastKind::Err,
                     }),
@@ -377,286 +231,31 @@ pub fn app() -> Html {
         })
     };
 
-    let on_new = {
-        let state = state.clone();
-        Callback::from(move |_: MouseEvent| state.dispatch(AppAction::GoToCreate))
-    };
-
-    let on_cancel_create = {
-        let state = state.clone();
-        Callback::from(move |_: ()| state.dispatch(AppAction::GoToList))
-    };
-
-    let on_create = {
-        let state = state.clone();
-        Callback::from(move |req: CreateRequest| {
-            let state = state.clone();
-            spawn_local(async move {
-                match api_create(&req).await {
-                    Ok(job) => {
-                        state.dispatch(AppAction::UpsertJob(job));
-                        state.dispatch(AppAction::GoToList);
-                        state.dispatch(AppAction::AddToast {
-                            msg: "Job created".into(),
-                            kind: ToastKind::Ok,
-                        });
-                    }
-                    Err(e) => state.dispatch(AppAction::AddToast {
-                        msg: format!("Create failed: {e}"),
-                        kind: ToastKind::Err,
-                    }),
-                }
-            })
+    let switch = {
+        let on_toast = on_toast.clone();
+        Callback::from(move |route: Route| match route {
+            Route::Home => html! { <Redirect<Route> to={Route::CronJobs} /> },
+            Route::CronJobs => html! { <CronJobsPage on_toast={on_toast.clone()} /> },
+            Route::Routines => html! { <RoutinesPage on_toast={on_toast.clone()} /> },
+            Route::NotFound => html! { <Redirect<Route> to={Route::CronJobs} /> },
         })
     };
 
-    let on_logs = {
-        let state = state.clone();
-        Callback::from(move |id: String| state.dispatch(AppAction::GoToLogs(id)))
-    };
-
-    let on_back_from_logs = {
-        let state = state.clone();
-        Callback::from(move |_: ()| state.dispatch(AppAction::GoToList))
-    };
-
-    let on_edit = {
-        let state = state.clone();
-        Callback::from(move |id: String| state.dispatch(AppAction::OpenEdit(id)))
-    };
-
-    let on_ask_delete = {
-        let state = state.clone();
-        Callback::from(move |(id, handler): (String, String)| {
-            state.dispatch(AppAction::OpenConfirmDelete { id, handler });
-        })
-    };
-
-    let on_trigger = {
-        let state = state.clone();
-        Callback::from(move |id: String| {
-            let state = state.clone();
-            spawn_local(async move {
-                match api_trigger(&id).await {
-                    Ok(job) => {
-                        state.dispatch(AppAction::UpsertJob(job));
-                        state.dispatch(AppAction::AddToast {
-                            msg: "Job triggered".into(),
-                            kind: ToastKind::Ok,
-                        });
-                    }
-                    Err(e) => state.dispatch(AppAction::AddToast {
-                        msg: format!("Trigger failed: {e}"),
-                        kind: ToastKind::Err,
-                    }),
-                }
-            })
-        })
-    };
-
-    let on_toggle = {
-        let state = state.clone();
-        Callback::from(move |(id, enabled): (String, bool)| {
-            let state = state.clone();
-            spawn_local(async move {
-                let req = UpdateRequest {
-                    enabled: Some(enabled),
-                    ..Default::default()
-                };
-                match api_update(&id, &req).await {
-                    Ok(job) => {
-                        state.dispatch(AppAction::UpsertJob(job));
-                        state.dispatch(AppAction::AddToast {
-                            msg: if enabled {
-                                "Job enabled".into()
-                            } else {
-                                "Job disabled".into()
-                            },
-                            kind: ToastKind::Ok,
-                        });
-                    }
-                    Err(e) => state.dispatch(AppAction::AddToast {
-                        msg: format!("Toggle failed: {e}"),
-                        kind: ToastKind::Err,
-                    }),
-                }
-            })
-        })
-    };
-
-    let on_close = {
-        let state = state.clone();
-        Callback::from(move |_: ()| state.dispatch(AppAction::CloseModal))
-    };
-
-    let current_modal = state.modal.clone();
-    let on_save = {
-        let state = state.clone();
-        Callback::from(move |req: CreateRequest| {
-            let state = state.clone();
-            let modal = current_modal.clone();
-            spawn_local(async move {
-                if let Modal::Edit(id) = &modal {
-                    let upd = UpdateRequest {
-                        schedule: Some(req.schedule),
-                        handler: Some(req.handler),
-                        metadata: Some(req.metadata),
-                        enabled: Some(req.enabled),
-                    };
-                    match api_update(id, &upd).await {
-                        Ok(job) => {
-                            state.dispatch(AppAction::UpsertJob(job));
-                            state.dispatch(AppAction::CloseModal);
-                            state.dispatch(AppAction::AddToast {
-                                msg: "Job updated".into(),
-                                kind: ToastKind::Ok,
-                            });
-                        }
-                        Err(e) => state.dispatch(AppAction::AddToast {
-                            msg: format!("Update failed: {e}"),
-                            kind: ToastKind::Err,
-                        }),
-                    }
-                }
-            })
-        })
-    };
-
-    let on_confirm_delete = {
-        let state = state.clone();
-        Callback::from(move |id: String| {
-            let state = state.clone();
-            spawn_local(async move {
-                match api_delete(&id).await {
-                    Ok(()) => {
-                        state.dispatch(AppAction::RemoveJob(id));
-                        state.dispatch(AppAction::CloseModal);
-                        state.dispatch(AppAction::AddToast {
-                            msg: "Job deleted".into(),
-                            kind: ToastKind::Ok,
-                        });
-                    }
-                    Err(e) => state.dispatch(AppAction::AddToast {
-                        msg: format!("Delete failed: {e}"),
-                        kind: ToastKind::Err,
-                    }),
-                }
-            })
-        })
-    };
-
-    let on_dismiss_toast = {
-        let state = state.clone();
-        Callback::from(move |id: u32| state.dispatch(AppAction::DismissToast(id)))
-    };
-
-    let on_tab = {
-        let state = state.clone();
-        Callback::from(move |tab: Tab| state.dispatch(AppAction::SwitchTab(tab)))
-    };
-
-    let on_routine_toast = {
-        let state = state.clone();
-        Callback::from(move |(msg, kind): (String, ToastKind)| {
-            state.dispatch(AppAction::AddToast { msg, kind })
-        })
-    };
-
-    let jobs = state.jobs.clone();
     let health = state.health.clone();
     let health_ok = state.health_ok;
-    let loading = state.loading;
-    let tab = state.tab;
-    let page = state.page.clone();
-    let modal = state.modal.clone();
     let toasts = state.toasts.clone();
-
-    let edit_job = match &modal {
-        Modal::Edit(id) => jobs.iter().find(|j| j.id == *id).cloned(),
-        _ => None,
-    };
-
-    let jobs_body = html! {
-        <>
-            {
-                match page {
-                    Page::NewJob => html! {
-                        <CreatePage on_cancel={on_cancel_create} on_save={on_create} />
-                    },
-                    Page::Logs(id) => {
-                        let handler = jobs.iter()
-                            .find(|j| j.id == id)
-                            .map(|j| j.handler.clone())
-                            .unwrap_or_default();
-                        html! {
-                            <LogsPage job_id={id} handler={handler} on_back={on_back_from_logs} />
-                        }
-                    },
-                    Page::List => html! {
-                        <main>
-                            <StatsBar jobs={jobs.clone()} />
-                            <div class="section-hd">
-                                <div class="section-label">{"SCHEDULED JOBS"}</div>
-                                <button class="btn btn-primary btn-sm" onclick={on_new}>{"+ NEW JOB"}</button>
-                            </div>
-                            <JobTable
-                                jobs={jobs}
-                                loading={loading}
-                                on_edit={on_edit}
-                                on_delete={on_ask_delete}
-                                on_toggle={on_toggle}
-                                on_trigger={on_trigger}
-                                on_logs={on_logs}
-                            />
-                        </main>
-                    },
-                }
-            }
-            {
-                match &modal {
-                    Modal::Edit(_) => html! {
-                        <JobModal
-                            editing={edit_job}
-                            on_close={on_close.clone()}
-                            on_save={on_save}
-                        />
-                    },
-                    Modal::ConfirmDelete { id, handler } => html! {
-                        <ConfirmDialog
-                            job_id={id.clone()}
-                            handler={handler.clone()}
-                            on_cancel={on_close.clone()}
-                            on_confirm={on_confirm_delete}
-                        />
-                    },
-                    // Shutdown is global (the STOP button lives in the always-visible
-                    // header), so its dialog is rendered at the top level, not here.
-                    Modal::ConfirmShutdown => html! {},
-                    Modal::None => html! {},
-                }
-            }
-        </>
-    };
+    let show_shutdown = state.show_shutdown;
 
     html! {
         <>
             <Header health={health} ok={health_ok} on_refresh={on_refresh} on_stop={on_stop} />
-            <TabNav active={tab} on_tab={on_tab} />
+            <Nav />
+            <Switch<Route> render={switch} />
             {
-                match tab {
-                    Tab::Jobs => jobs_body,
-                    Tab::Routines => html! {
-                        <RoutinesPage on_toast={on_routine_toast} />
-                    },
-                }
-            }
-            // Global shutdown confirmation — rendered here (not inside a tab) so it
-            // appears no matter which tab is active when STOP is pressed.
-            {
-                if modal == Modal::ConfirmShutdown {
+                if show_shutdown {
                     html! {
                         <ShutdownDialog
-                            on_cancel={on_close}
+                            on_cancel={on_close_shutdown}
                             on_confirm={on_confirm_shutdown}
                         />
                     }
@@ -664,56 +263,34 @@ pub fn app() -> Html {
                     html! {}
                 }
             }
-            <ToastStack toasts={toasts} on_dismiss={on_dismiss_toast} />
+            <ToastStack toasts={toasts} />
         </>
     }
 }
 
-// ─── Tab nav ──────────────────────────────────────────────────────────────────
+// ─── Nav ──────────────────────────────────────────────────────────────────────
 
-#[derive(Properties, PartialEq)]
-pub struct TabNavProps {
-    pub active: Tab,
-    pub on_tab: Callback<Tab>,
-}
-
-#[function_component(TabNav)]
-pub fn tab_nav(props: &TabNavProps) -> Html {
-    let mk = |tab: Tab, label: &'static str| {
-        let cb = props.on_tab.clone();
-        let cls = if props.active == tab {
+#[function_component(Nav)]
+pub fn nav() -> Html {
+    let route = use_route::<Route>().unwrap_or(Route::Home);
+    // Home redirects to CronJobs, so treat it as the cron-jobs tab for highlighting.
+    let cls = |target: &Route| {
+        let active = route == *target || (route == Route::Home && *target == Route::CronJobs);
+        if active {
             "tab-btn active"
         } else {
             "tab-btn"
-        };
-        html! {
-            <button class={cls} onclick={Callback::from(move |_: MouseEvent| cb.emit(tab))}>
-                { label }
-            </button>
         }
     };
     html! {
         <nav class="tabs">
-            { mk(Tab::Jobs, "CRON JOBS") }
-            { mk(Tab::Routines, "ROUTINES") }
+            <Link<Route> classes={classes!(cls(&Route::CronJobs))} to={Route::CronJobs}>
+                { "CRON JOBS" }
+            </Link<Route>>
+            <Link<Route> classes={classes!(cls(&Route::Routines))} to={Route::Routines}>
+                { "ROUTINES" }
+            </Link<Route>>
         </nav>
-    }
-}
-
-async fn poll_health(state: UseReducerHandle<AppState>) {
-    match api_health().await {
-        Ok(health) => {
-            let ok = health.running;
-            state.dispatch(AppAction::HealthLoaded { health, ok });
-        }
-        Err(_) => state.dispatch(AppAction::HealthLoaded {
-            health: Health {
-                status: "offline".into(),
-                running: false,
-                uptime_secs: None,
-            },
-            ok: false,
-        }),
     }
 }
 
@@ -745,7 +322,7 @@ pub fn header(props: &HeaderProps) -> Html {
         <header>
             <div class="logo">
                 {"MOADIM"}
-                <span class="logo-sub">{"/ CRON CONTROL"}</span>
+                <span class="logo-sub">{"/ CONTROL"}</span>
             </div>
             <div class="header-right">
                 <div class="health">
@@ -757,652 +334,6 @@ pub fn header(props: &HeaderProps) -> Html {
                 <button class="btn-stop" title="Stop the server" disabled={!props.ok} onclick={props.on_stop.clone()}>{"⏻ STOP"}</button>
             </div>
         </header>
-    }
-}
-
-// ─── Stats bar ────────────────────────────────────────────────────────────────
-
-#[derive(Properties, PartialEq)]
-pub struct StatsProps {
-    pub jobs: Vec<CronJob>,
-}
-
-#[function_component(StatsBar)]
-pub fn stats_bar(props: &StatsProps) -> Html {
-    let total = props.jobs.len();
-    let enabled = props.jobs.iter().filter(|j| j.enabled).count();
-    let disabled = total - enabled;
-
-    html! {
-        <div class="stats">
-            <div class="stat-card all">
-                <div class="stat-label">{"TOTAL JOBS"}</div>
-                <div class="stat-val">{total}</div>
-            </div>
-            <div class="stat-card enabled">
-                <div class="stat-label">{"ENABLED"}</div>
-                <div class="stat-val c-accent">{enabled}</div>
-            </div>
-            <div class="stat-card disabled">
-                <div class="stat-label">{"DISABLED"}</div>
-                <div class="stat-val c-amber">{disabled}</div>
-            </div>
-        </div>
-    }
-}
-
-// ─── Job table ────────────────────────────────────────────────────────────────
-
-#[derive(Properties, PartialEq)]
-pub struct JobTableProps {
-    pub jobs: Vec<CronJob>,
-    pub loading: bool,
-    pub on_edit: Callback<String>,
-    pub on_delete: Callback<(String, String)>,
-    pub on_toggle: Callback<(String, bool)>,
-    pub on_trigger: Callback<String>,
-    pub on_logs: Callback<String>,
-}
-
-#[function_component(JobTable)]
-pub fn job_table(props: &JobTableProps) -> Html {
-    if props.loading {
-        return html! {
-            <div class="table-wrap">
-                <div class="empty"><div class="spinner"></div></div>
-            </div>
-        };
-    }
-    if props.jobs.is_empty() {
-        return html! {
-            <div class="table-wrap">
-                <div class="empty">
-                    <div class="empty-icon">{"⧗"}</div>
-                    <div class="empty-msg">{"NO JOBS SCHEDULED"}</div>
-                    <div class="empty-sub">{"press + NEW JOB to create one"}</div>
-                </div>
-            </div>
-        };
-    }
-
-    html! {
-        <div class="table-wrap">
-            <table>
-                <thead>
-                    <tr>
-                        <th>{"ID"}</th>
-                        <th>{"SCHEDULE"}</th>
-                        <th>{"HANDLER"}</th>
-                        <th>{"METADATA"}</th>
-                        <th>{"ENABLED"}</th>
-                        <th>{"UPDATED"}</th>
-                        <th></th>
-                    </tr>
-                </thead>
-                <tbody>
-                    { for props.jobs.iter().map(|job| html! {
-                        <JobRow
-                            key={job.id.clone()}
-                            job={job.clone()}
-                            on_edit={props.on_edit.clone()}
-                            on_delete={props.on_delete.clone()}
-                            on_toggle={props.on_toggle.clone()}
-                            on_trigger={props.on_trigger.clone()}
-                            on_logs={props.on_logs.clone()}
-                        />
-                    }) }
-                </tbody>
-            </table>
-        </div>
-    }
-}
-
-// ─── Job row ──────────────────────────────────────────────────────────────────
-
-#[derive(Properties, PartialEq)]
-pub struct JobRowProps {
-    pub job: CronJob,
-    pub on_edit: Callback<String>,
-    pub on_delete: Callback<(String, String)>,
-    pub on_toggle: Callback<(String, bool)>,
-    pub on_trigger: Callback<String>,
-    pub on_logs: Callback<String>,
-}
-
-#[function_component(JobRow)]
-pub fn job_row(props: &JobRowProps) -> Html {
-    let job = &props.job;
-    let id_short = format!("{}…", &job.id[..8.min(job.id.len())]);
-    let cron_text = job
-        .schedule_description
-        .as_deref()
-        .unwrap_or("—")
-        .to_string();
-    let meta = meta_preview(&job.metadata);
-    let updated = reltime(job.updated_at);
-
-    let on_edit = {
-        let cb = props.on_edit.clone();
-        let id = job.id.clone();
-        Callback::from(move |_: MouseEvent| cb.emit(id.clone()))
-    };
-    let on_delete = {
-        let cb = props.on_delete.clone();
-        let id = job.id.clone();
-        let handler = job.handler.clone();
-        Callback::from(move |_: MouseEvent| cb.emit((id.clone(), handler.clone())))
-    };
-    let on_toggle = {
-        let cb = props.on_toggle.clone();
-        let id = job.id.clone();
-        let enabled = job.enabled;
-        Callback::from(move |_: Event| cb.emit((id.clone(), !enabled)))
-    };
-    let on_trigger = {
-        let cb = props.on_trigger.clone();
-        let id = job.id.clone();
-        Callback::from(move |_: MouseEvent| cb.emit(id.clone()))
-    };
-    let on_logs = {
-        let cb = props.on_logs.clone();
-        let id = job.id.clone();
-        Callback::from(move |_: MouseEvent| cb.emit(id.clone()))
-    };
-
-    let last_run = job
-        .last_triggered_at
-        .map(|t| format!("↻ {}", reltime(t)))
-        .unwrap_or_default();
-
-    html! {
-        <tr>
-            <td><span class="cell-id" title={job.id.clone()}>{id_short}</span></td>
-            <td>
-                <div class="cell-schedule">{&job.schedule}</div>
-                <div class="cell-schedule-human">{cron_text}</div>
-            </td>
-            <td><span class="cell-handler">{&job.handler}</span></td>
-            <td><span class="cell-meta">{meta}</span></td>
-            <td>
-                <label class="toggle">
-                    <input type="checkbox" checked={job.enabled} onchange={on_toggle} />
-                    <div class="toggle-track"></div>
-                </label>
-            </td>
-            <td>
-                <div class="cell-time">{updated}</div>
-                if !last_run.is_empty() {
-                    <div class="cell-triggered">{last_run}</div>
-                }
-            </td>
-            <td>
-                <div class="row-actions">
-                    <button class="act-btn run" title="Run now" onclick={on_trigger}>{"▶"}</button>
-                    <button class="act-btn logs" onclick={on_logs}>{"LOGS"}</button>
-                    <button class="act-btn edit" onclick={on_edit}>{"EDIT"}</button>
-                    <button class="act-btn del" onclick={on_delete}>{"✕"}</button>
-                </div>
-            </td>
-        </tr>
-    }
-}
-
-// ─── Create page ──────────────────────────────────────────────────────────────
-
-#[derive(Properties, PartialEq)]
-pub struct CreatePageProps {
-    pub on_cancel: Callback<()>,
-    pub on_save: Callback<CreateRequest>,
-}
-
-#[function_component(CreatePage)]
-pub fn create_page(props: &CreatePageProps) -> Html {
-    let schedule = use_state(String::new);
-    let handler = use_state(String::new);
-    let meta_raw = use_state(String::new);
-    let enabled = use_state(|| true);
-    let meta_err = use_state(String::new);
-    let saving = use_state(|| false);
-
-    let (cron_ok, cron_text) = describe_cron_live(&schedule);
-
-    let on_schedule = {
-        let schedule = schedule.clone();
-        Callback::from(move |e: InputEvent| {
-            let input: HtmlInputElement = e.target_unchecked_into();
-            schedule.set(input.value());
-        })
-    };
-    let on_handler = {
-        let handler = handler.clone();
-        Callback::from(move |e: InputEvent| {
-            let input: HtmlInputElement = e.target_unchecked_into();
-            handler.set(input.value());
-        })
-    };
-    let on_meta = {
-        let meta_raw = meta_raw.clone();
-        let meta_err = meta_err.clone();
-        Callback::from(move |e: InputEvent| {
-            let input: HtmlInputElement = e.target_unchecked_into();
-            let val = input.value();
-            if val.trim().is_empty() {
-                meta_err.set(String::new());
-            } else if let Err(err) = serde_json::from_str::<Json>(&val) {
-                meta_err.set(format!("↳ {err}"));
-            } else {
-                meta_err.set(String::new());
-            }
-            meta_raw.set(val);
-        })
-    };
-    let on_enabled = {
-        let enabled = enabled.clone();
-        Callback::from(move |e: Event| {
-            let input: HtmlInputElement = e.target_unchecked_into();
-            enabled.set(input.checked());
-        })
-    };
-
-    let set_preset = |val: &'static str| {
-        let schedule = schedule.clone();
-        Callback::from(move |_: MouseEvent| schedule.set(val.to_string()))
-    };
-
-    let on_cancel_click = {
-        let cb = props.on_cancel.clone();
-        Callback::from(move |_: MouseEvent| cb.emit(()))
-    };
-    let on_save_click = {
-        let schedule = schedule.clone();
-        let handler = handler.clone();
-        let meta_raw = meta_raw.clone();
-        let meta_err = meta_err.clone();
-        let enabled = enabled.clone();
-        let saving = saving.clone();
-        let cb = props.on_save.clone();
-        Callback::from(move |_: MouseEvent| {
-            if !meta_err.is_empty() {
-                return;
-            }
-            let metadata = if meta_raw.trim().is_empty() {
-                Json::Null
-            } else {
-                serde_json::from_str(&meta_raw).unwrap_or(Json::Null)
-            };
-            saving.set(true);
-            cb.emit(CreateRequest {
-                schedule: (*schedule).clone(),
-                handler: (*handler).clone(),
-                metadata,
-                enabled: *enabled,
-            });
-        })
-    };
-
-    let preview_class = if schedule.is_empty() {
-        "cron-preview"
-    } else if cron_ok {
-        "cron-preview ok"
-    } else {
-        "cron-preview bad"
-    };
-
-    let meta_class = if !meta_err.is_empty() {
-        "form-input invalid"
-    } else {
-        "form-input"
-    };
-
-    html! {
-        <main class="create-page">
-            <div class="page-hd">
-                <button class="btn btn-ghost btn-sm" onclick={on_cancel_click.clone()}>{"← BACK"}</button>
-                <div class="page-title">{"NEW JOB"}</div>
-            </div>
-            <div class="page-card">
-                <div class="modal-body">
-                    <div class="form-group">
-                        <label class="form-label">
-                            {"SCHEDULE "}
-                            <span class="form-required">{"*"}</span>
-                        </label>
-                        <input
-                            class="form-input"
-                            type="text"
-                            placeholder="sec min hour dom month dow year"
-                            value={(*schedule).clone()}
-                            oninput={on_schedule}
-                            autocomplete="off"
-                            spellcheck="false"
-                        />
-                        <div class="cron-presets">
-                            { for [
-                                ("@daily", "@daily"), ("@hourly", "@hourly"),
-                                ("@weekly", "@weekly"), ("@monthly", "@monthly"),
-                                ("0 0 9 * * 1-5 *", "weekdays 9am"),
-                                ("0 */15 * * * * *", "every 15min"),
-                                ("0 0 * * * * *", "every hour"),
-                                ("0 0 0 1 * * *", "monthly"),
-                            ].iter().map(|(val, label)| html! {
-                                <button class="preset-btn" onclick={set_preset(val)}>{*label}</button>
-                            }) }
-                        </div>
-                        <div class={preview_class}>{cron_text}</div>
-                    </div>
-                    <div class="form-group">
-                        <label class="form-label">
-                            {"HANDLER "}
-                            <span class="form-required">{"*"}</span>
-                        </label>
-                        <input
-                            class="form-input"
-                            type="text"
-                            placeholder="send-report"
-                            value={(*handler).clone()}
-                            oninput={on_handler}
-                            autocomplete="off"
-                            spellcheck="false"
-                        />
-                    </div>
-                    <div class="form-group">
-                        <label class="form-label">
-                            {"METADATA "}
-                            <span style="color:var(--text-ghost)">{"(JSON)"}</span>
-                        </label>
-                        <textarea
-                            class={meta_class}
-                            placeholder={r#"{"recipient": "team@example.com"}"#}
-                            value={(*meta_raw).clone()}
-                            oninput={on_meta}
-                        />
-                        if !meta_err.is_empty() {
-                            <div class="field-err">{(*meta_err).clone()}</div>
-                        }
-                    </div>
-                    <div class="form-group" style="margin-bottom:0">
-                        <div class="toggle-row">
-                            <span class="toggle-row-label">{"ENABLED"}</span>
-                            <label class="toggle">
-                                <input type="checkbox" checked={*enabled} onchange={on_enabled} />
-                                <div class="toggle-track"></div>
-                            </label>
-                        </div>
-                    </div>
-                </div>
-                <div class="modal-ft">
-                    <button class="btn btn-ghost btn-sm" onclick={on_cancel_click}>{"CANCEL"}</button>
-                    <button
-                        class="btn btn-primary btn-sm"
-                        onclick={on_save_click}
-                        disabled={*saving}
-                    >
-                        { if *saving { "…" } else { "CREATE JOB" } }
-                    </button>
-                </div>
-            </div>
-        </main>
-    }
-}
-
-// ─── Job modal (edit only) ────────────────────────────────────────────────────
-
-#[derive(Properties, PartialEq)]
-pub struct JobModalProps {
-    pub editing: Option<CronJob>,
-    pub on_close: Callback<()>,
-    pub on_save: Callback<CreateRequest>,
-}
-
-#[function_component(JobModal)]
-pub fn job_modal(props: &JobModalProps) -> Html {
-    let schedule = use_state(|| {
-        props
-            .editing
-            .as_ref()
-            .map(|j| j.schedule.clone())
-            .unwrap_or_default()
-    });
-    let handler = use_state(|| {
-        props
-            .editing
-            .as_ref()
-            .map(|j| j.handler.clone())
-            .unwrap_or_default()
-    });
-    let meta_raw = use_state(|| {
-        props
-            .editing
-            .as_ref()
-            .and_then(|j| {
-                if j.metadata.is_null() {
-                    None
-                } else {
-                    serde_json::to_string_pretty(&j.metadata).ok()
-                }
-            })
-            .unwrap_or_default()
-    });
-    let enabled = use_state(|| props.editing.as_ref().map(|j| j.enabled).unwrap_or(true));
-    let meta_err = use_state(String::new);
-    let saving = use_state(|| false);
-
-    let (cron_ok, cron_text) = describe_cron_live(&schedule);
-
-    let on_schedule = {
-        let schedule = schedule.clone();
-        Callback::from(move |e: InputEvent| {
-            let input: HtmlInputElement = e.target_unchecked_into();
-            schedule.set(input.value());
-        })
-    };
-    let on_handler = {
-        let handler = handler.clone();
-        Callback::from(move |e: InputEvent| {
-            let input: HtmlInputElement = e.target_unchecked_into();
-            handler.set(input.value());
-        })
-    };
-    let on_meta = {
-        let meta_raw = meta_raw.clone();
-        let meta_err = meta_err.clone();
-        Callback::from(move |e: InputEvent| {
-            let input: HtmlInputElement = e.target_unchecked_into();
-            let val = input.value();
-            if val.trim().is_empty() {
-                meta_err.set(String::new());
-            } else if let Err(err) = serde_json::from_str::<Json>(&val) {
-                meta_err.set(format!("↳ {err}"));
-            } else {
-                meta_err.set(String::new());
-            }
-            meta_raw.set(val);
-        })
-    };
-    let on_enabled = {
-        let enabled = enabled.clone();
-        Callback::from(move |e: Event| {
-            let input: HtmlInputElement = e.target_unchecked_into();
-            enabled.set(input.checked());
-        })
-    };
-
-    let set_preset = |val: &'static str| {
-        let schedule = schedule.clone();
-        Callback::from(move |_: MouseEvent| schedule.set(val.to_string()))
-    };
-
-    let on_close_click = {
-        let cb = props.on_close.clone();
-        Callback::from(move |_: MouseEvent| cb.emit(()))
-    };
-    let on_save_click = {
-        let schedule = schedule.clone();
-        let handler = handler.clone();
-        let meta_raw = meta_raw.clone();
-        let meta_err = meta_err.clone();
-        let enabled = enabled.clone();
-        let saving = saving.clone();
-        let cb = props.on_save.clone();
-        Callback::from(move |_: MouseEvent| {
-            if !meta_err.is_empty() {
-                return;
-            }
-            let metadata = if meta_raw.trim().is_empty() {
-                Json::Null
-            } else {
-                serde_json::from_str(&meta_raw).unwrap_or(Json::Null)
-            };
-            saving.set(true);
-            cb.emit(CreateRequest {
-                schedule: (*schedule).clone(),
-                handler: (*handler).clone(),
-                metadata,
-                enabled: *enabled,
-            });
-        })
-    };
-
-    let preview_class = if schedule.is_empty() {
-        "cron-preview"
-    } else if cron_ok {
-        "cron-preview ok"
-    } else {
-        "cron-preview bad"
-    };
-
-    let meta_class = if !meta_err.is_empty() {
-        "form-input invalid"
-    } else {
-        "form-input"
-    };
-
-    html! {
-        <div class="overlay open">
-            <div class="modal">
-                <div class="modal-hd">
-                    <div class="modal-title">{"EDIT JOB"}</div>
-                    <button class="modal-x" onclick={on_close_click.clone()}>{"✕"}</button>
-                </div>
-                <div class="modal-body">
-                    <div class="form-group">
-                        <label class="form-label">
-                            {"SCHEDULE "}
-                            <span class="form-required">{"*"}</span>
-                        </label>
-                        <input
-                            class="form-input"
-                            type="text"
-                            placeholder="sec min hour dom month dow year"
-                            value={(*schedule).clone()}
-                            oninput={on_schedule}
-                            autocomplete="off"
-                            spellcheck="false"
-                        />
-                        <div class="cron-presets">
-                            { for [
-                                ("@daily", "@daily"), ("@hourly", "@hourly"),
-                                ("@weekly", "@weekly"), ("@monthly", "@monthly"),
-                                ("0 0 9 * * 1-5 *", "weekdays 9am"),
-                                ("0 */15 * * * * *", "every 15min"),
-                                ("0 0 * * * * *", "every hour"),
-                                ("0 0 0 1 * * *", "monthly"),
-                            ].iter().map(|(val, label)| html! {
-                                <button class="preset-btn" onclick={set_preset(val)}>{*label}</button>
-                            }) }
-                        </div>
-                        <div class={preview_class}>{cron_text}</div>
-                    </div>
-                    <div class="form-group">
-                        <label class="form-label">
-                            {"HANDLER "}
-                            <span class="form-required">{"*"}</span>
-                        </label>
-                        <input
-                            class="form-input"
-                            type="text"
-                            placeholder="send-report"
-                            value={(*handler).clone()}
-                            oninput={on_handler}
-                            autocomplete="off"
-                            spellcheck="false"
-                        />
-                    </div>
-                    <div class="form-group">
-                        <label class="form-label">
-                            {"METADATA "}
-                            <span style="color:var(--text-ghost)">{"(JSON)"}</span>
-                        </label>
-                        <textarea
-                            class={meta_class}
-                            placeholder={r#"{"recipient": "team@example.com"}"#}
-                            value={(*meta_raw).clone()}
-                            oninput={on_meta}
-                        />
-                        if !meta_err.is_empty() {
-                            <div class="field-err">{(*meta_err).clone()}</div>
-                        }
-                    </div>
-                    <div class="form-group" style="margin-bottom:0">
-                        <div class="toggle-row">
-                            <span class="toggle-row-label">{"ENABLED"}</span>
-                            <label class="toggle">
-                                <input type="checkbox" checked={*enabled} onchange={on_enabled} />
-                                <div class="toggle-track"></div>
-                            </label>
-                        </div>
-                    </div>
-                </div>
-                <div class="modal-ft">
-                    <button class="btn btn-ghost btn-sm" onclick={on_close_click}>{"CANCEL"}</button>
-                    <button
-                        class="btn btn-primary btn-sm"
-                        onclick={on_save_click}
-                        disabled={*saving}
-                    >
-                        { if *saving { "…" } else { "SAVE CHANGES" } }
-                    </button>
-                </div>
-            </div>
-        </div>
-    }
-}
-
-// ─── Confirm dialog ───────────────────────────────────────────────────────────
-
-#[derive(Properties, PartialEq)]
-pub struct ConfirmProps {
-    pub job_id: String,
-    pub handler: String,
-    pub on_cancel: Callback<()>,
-    pub on_confirm: Callback<String>,
-}
-
-#[function_component(ConfirmDialog)]
-pub fn confirm_dialog(props: &ConfirmProps) -> Html {
-    let on_cancel = {
-        let cb = props.on_cancel.clone();
-        Callback::from(move |_: MouseEvent| cb.emit(()))
-    };
-    let on_confirm = {
-        let cb = props.on_confirm.clone();
-        let id = props.job_id.clone();
-        Callback::from(move |_: MouseEvent| cb.emit(id.clone()))
-    };
-
-    html! {
-        <div class="overlay open">
-            <div class="confirm-dialog">
-                <div class="confirm-title">{"⚠ DELETE JOB"}</div>
-                <div class="confirm-msg">
-                    { format!("Delete the job running \"{}\"? This cannot be undone.", props.handler) }
-                </div>
-                <div class="confirm-acts">
-                    <button class="btn btn-ghost btn-sm" onclick={on_cancel}>{"CANCEL"}</button>
-                    <button class="btn btn-danger btn-sm" onclick={on_confirm}>{"DELETE"}</button>
-                </div>
-            </div>
-        </div>
     }
 }
 
@@ -1446,7 +377,6 @@ pub fn shutdown_dialog(props: &ShutdownProps) -> Html {
 #[derive(Properties, PartialEq)]
 pub struct ToastStackProps {
     pub toasts: Vec<Toast>,
-    pub on_dismiss: Callback<u32>,
 }
 
 #[function_component(ToastStack)]
@@ -1463,106 +393,7 @@ pub fn toast_stack(props: &ToastStackProps) -> Html {
     }
 }
 
-// ─── Logs page ────────────────────────────────────────────────────────────────
-
-#[derive(Properties, PartialEq)]
-pub struct LogsPageProps {
-    pub job_id: String,
-    pub handler: String,
-    pub on_back: Callback<()>,
-}
-
-#[function_component(LogsPage)]
-pub fn logs_page(props: &LogsPageProps) -> Html {
-    let content: UseStateHandle<Option<String>> = use_state(|| None);
-    let loading = use_state(|| true);
-    let err: UseStateHandle<Option<String>> = use_state(|| None);
-
-    {
-        let id = props.job_id.clone();
-        let content = content.clone();
-        let loading = loading.clone();
-        let err = err.clone();
-        use_effect_with(id.clone(), move |id| {
-            let id = id.clone();
-            let content = content.clone();
-            let loading = loading.clone();
-            let err = err.clone();
-            spawn_local(async move {
-                match api_get_logs(&id).await {
-                    Ok(text) => {
-                        content.set(Some(text));
-                        err.set(None);
-                    }
-                    Err(e) => err.set(Some(e)),
-                }
-                loading.set(false);
-            });
-        });
-    }
-
-    let on_back = {
-        let cb = props.on_back.clone();
-        Callback::from(move |_: MouseEvent| cb.emit(()))
-    };
-
-    let on_refresh = {
-        let id = props.job_id.clone();
-        let content = content.clone();
-        let loading = loading.clone();
-        let err = err.clone();
-        Callback::from(move |_: MouseEvent| {
-            let id = id.clone();
-            let content = content.clone();
-            let loading = loading.clone();
-            let err = err.clone();
-            loading.set(true);
-            spawn_local(async move {
-                match api_get_logs(&id).await {
-                    Ok(text) => {
-                        content.set(Some(text));
-                        err.set(None);
-                    }
-                    Err(e) => err.set(Some(e)),
-                }
-                loading.set(false);
-            });
-        })
-    };
-
-    let is_loading = *loading;
-    let err_msg = (*err).clone();
-    let log_text = (*content).clone();
-
-    let body = if is_loading {
-        html! { <div class="empty"><div class="spinner"></div></div> }
-    } else if let Some(e) = err_msg {
-        html! { <div class="logs-error">{format!("Error: {e}")}</div> }
-    } else if let Some(text) = log_text {
-        if text.is_empty() {
-            html! { <div class="logs-empty">{"— no logs yet —"}</div> }
-        } else {
-            html! { <pre class="logs-content">{text}</pre> }
-        }
-    } else {
-        html! {}
-    };
-
-    html! {
-        <main class="logs-page">
-            <div class="page-hd">
-                <button class="btn btn-ghost btn-sm" onclick={on_back}>{"← BACK"}</button>
-                <div class="page-title">{format!("LOGS / {}", props.handler)}</div>
-                <button class="btn-refresh" title="Refresh" onclick={on_refresh}>{"↻"}</button>
-            </div>
-            <div class="logs-wrap">
-                {body}
-            </div>
-        </main>
-    }
-}
-
-// ─── Utilities ────────────────────────────────────────────────────────────────
+// ─── Utilities (shared with routines + cron_jobs modules) ─────────────────────
 
 /// Parse a cron expression into a `Cron`, normalizing the 7-field
 /// (sec min hour dom month dow year) form to 5-field to match server behaviour.
@@ -1593,22 +424,6 @@ pub(crate) fn describe_cron_live(expr: &str) -> (bool, String) {
     match parse_cron(expr) {
         Some(cron) => (true, cron.describe()),
         None => (false, "Invalid cron expression".into()),
-    }
-}
-
-fn meta_preview(v: &Json) -> String {
-    match v {
-        Json::Null => "—".into(),
-        Json::Object(o) if o.is_empty() => "{}".into(),
-        Json::Object(o) => {
-            let k = o.len();
-            if k == 1 {
-                format!("{{{}}}", o.keys().next().unwrap())
-            } else {
-                format!("{{{k} keys}}")
-            }
-        }
-        other => other.to_string().chars().take(24).collect(),
     }
 }
